Fix failing test after cloning repo for the first time.
Currently the tests will fail on a fresh checkout of the repo, but will
fail the second time the tests are run. This is because the last test
is not waiting until the I/O has finished in runFixture to make its
final assertion.

This fixes that issue by adding an optional callback to runFixture.