Only match a src attribute
resrcify is supposed to match src attributes, but currently, it matches
any attribute that ends in src, for instance ng-src. This makes it hard
to use other image sources. By adding a check to see if the src attribute
is following a space, a single or a doble quote, it still matches any src
tag, but not more than that.